doc: renaming algorithm main file
